requirements.txt
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncore-wsgi==0.0.11
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
bottle==0.12.25
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
deptry==0.20.0
Django==5.0.6
django-admin-interface==0.28.8
django-bootstrap-modal-forms==3.0.4
django-colorfield==0.11.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.6.2
django-organizations==2.4.1
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
fqdn==1.5.1
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importmagic3==0.2.0
ipykernel==6.29.5
ipython==8.21.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-console==6.5.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
log_colorizer==2.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.5.4
notebook_shim==0.2.4
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pipdeptree==2.23.1
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydot==3.0.1
Pygments==2.18.0
pyinotify==0.9.6
Pympler==1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.4
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spark-parser==1.8.9
sqlparse==0.5.0
stack-data==0.6.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tornado_systemd==1.0.1
traitlets==5.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uncompyle6==3.9.2
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
xdis==6.1.1
